adapt to prefix change
