+++ common/autoware_motion_utils/include/autoware/motion_utils/trajectory_v2/trajectory.hpp
@@ -1,0 +1,21 @@
+// Copyright 2024 Tier IV, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef AUTOWARE__MOTION_UTILS__TRAJECTORY_V2__TRAJECTORY_HPP_
+#define AUTOWARE__MOTION_UTILS__TRAJECTORY_V2__TRAJECTORY_HPP_
+#include "autoware/motion_utils/trajectory_v2/trajectory/trajectory_path_point.hpp"
+#include "autoware/motion_utils/trajectory_v2/trajectory/trajectory_path_point_with_lane_id.hpp"
+#include "autoware/motion_utils/trajectory_v2/trajectory/trajectory_point.hpp"
+#include "autoware/motion_utils/trajectory_v2/trajectory/trajectory_pose.hpp"
+#endif  // AUTOWARE__MOTION_UTILS__TRAJECTORY_V2__TRAJECTORY_HPP_
