--- common/motion_utils/include/motion_utils/trajectory_v2/interpolator.hpp
@@ -1,22 +1,0 @@
-// Copyright 2024 Tier IV, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#ifndef MOTION_UTILS__TRAJECTORY_V2__INTERPOLATOR_HPP_
-#define MOTION_UTILS__TRAJECTORY_V2__INTERPOLATOR_HPP_
-#include "motion_utils/trajectory_v2/interpolator/akima_spline.hpp"
-#include "motion_utils/trajectory_v2/interpolator/cubic_spline.hpp"
-#include "motion_utils/trajectory_v2/interpolator/linear.hpp"
-#include "motion_utils/trajectory_v2/interpolator/nearest_neighbor.hpp"
-#include "motion_utils/trajectory_v2/interpolator/zero_order_hold.hpp"
-#endif  // MOTION_UTILS__TRAJECTORY_V2__INTERPOLATOR_HPP_
